Refactor HistoryTable
